To do list - [Task 1]
Adding tasks: To add a task, simply type it in at the prompt and press enter. The task will be added to the to-do list.
Removing tasks: To remove a task, type the task name at the prompt and press enter. The task will be removed from the to-do list if it exists.
Listing tasks: To list your tasks, simply type 3 at the prompt and press enter. The app will print a list of all your tasks, numbered from 1.
Quitting the app: To quit the app, simply type 4 at the prompt and press enter.